move port onto output manager
