+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/promisehandlers/AssignPortHandler.scala
@@ -22,7 +22,7 @@ trait AssignPortHandler {
     if (msg.input) {
       dp.inputGateway.addPort(msg.portId, msg.schema)
     } else {
-      dp.outputGateway.addPort(msg.portId, msg.schema)
+      dp.outputManager.addPort(msg.portId, msg.schema)
     }
   }
